freqmeter/requirements.txt
numpy>=1.19.1
PyQt5>=5.15.0
PyQt5-sip>=12.8.1
pyqtgraph>=0.11.0
pyserial>=3.4
